% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTRL.R
\name{.calcEventRainfallMetrics}
\alias{.calcEventRainfallMetrics}
\title{calcEventRainfallMetrics}
\usage{
.calcEventRainfallMetrics(
  x,
  dates,
  list.RainEvents,
  modus,
  RD = RD,
  MAP = MAP,
  RDN = RDN
)
}
\arguments{
\item{x}{vector containing precipitation}

\item{dates}{see CTRL()}

\item{list.RainEvents}{list containing indices of rain events}

\item{modus}{modus of calculation of rain metrics. "sub" of sub-rainfall events or "main" for critical rainfall event}

\item{RD}{see CTRL()}

\item{MAP}{see CTRL()}

\item{RDN}{see CTRL()}
}
\value{
vector containing rainfall metrics
}
\description{
This function calcuates different precipitation metrics.
}
