% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot}
\alias{plot}
\title{plot}
\usage{
plot(
  x,
  y = NULL,
  col.pts = "darkblue",
  size.pts = 1,
  shape.pts = 16,
  col.line = "black",
  size.line = 0.8,
  col.line.ci = "darkgrey",
  size.line.ci = 0.6,
  linetype.ci = "dashed",
  fill.ci = "grey",
  alpha.ci = 0.4
)
}
\arguments{
\item{x}{object of class RainSlideThresh}

\item{col.pts}{colour of ED-pair points. Default: 'darkblue'}

\item{size.pts}{size of ED-pair points. Default: 1}

\item{shape.pts}{shape of ED-pair points. Default: 16}

\item{col.line}{colour of threshold line. Default: 'black'}

\item{size.line}{size of threshold line. Default: 0.8}

\item{col.line.ci}{colour of confidence interval line. Default: 'darkgrey'}

\item{size.line.ci}{size of confidence interval line. Default: 0.6}

\item{linetype.ci}{linetype of confidence interval. Default: 'dashed'}

\item{fill.ci}{inner fill of confidence interval. Default: 'grey'}

\item{alpha.ci}{transparency of confidence interval fill. Default: 0.4}
}
\description{
plot
}
\keyword{internal}
