% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh.R
\name{thresh}
\alias{thresh}
\title{Get rain event of a landslide}
\usage{
thresh(
  Re,
  D,
  method = c("LS", "QR", "NLS"),
  prob.thresh = 0.05,
  trans.log10 = FALSE,
  bootstrapping = TRUE,
  R = 1000,
  use.bootMedian = FALSE,
  use.normal = FALSE,
  nls.pw = 10,
  nls.tw = 10,
  nls.method = c("tw", "pw"),
  nls.bounds = list(lower = c(t = 0, alpha = 0, gamma = 0), upper = c(t = 500, alpha =
    100, gamma = 10)),
  seed = 123,
  cores = 1
)
}
\arguments{
\item{Re}{vector containing the rain event variable, e.g. cumulated event rainfall (in mm) or intensity (mm/h)}

\item{D}{vector containing the duration of the rainfall events}

\item{method}{method to compute threshold. Either 'LS' for least square, or 'NLS' for non-linear least squares Method. Default: 'nls'}

\item{prob.thresh}{exceedance probability level. Default: 0.05 (5 percent)}

\item{trans.log10}{log-transformation of input vectors Re and/or D. Default: TRUE}

\item{bootstrapping}{If TRUE bootstrapping is performed. Default: TRUE}

\item{R}{the number of bootstrap replicates, see boot::boot() for more information. Default: 1000}

\item{use.bootMedian}{Threshold is delineated from median of bootstrapping result. Default: FALSE}

\item{use.normal}{Use normal approximation of residuals This method is descriped in literature. Default: FALSE}

\item{nls.pw}{For NLS method, size of point-wise ED-pair filtering. Default: 10}

\item{nls.tw}{For NLS method, size of time-wise ED-pair filtering. Default: 10}

\item{nls.method}{Definde method for NLS ED-pair fitlering. Default: 'tw'}

\item{nls.bounds}{bounds for finding NLS starting parameter. Default: list(lower = c(t = 0, alpha = 0, gamma = 0), upper = c(t = 500, alpha = 100, gamma = 10))}

\item{seed}{replicable bootstrapping. Default: 123}

\item{cores}{If cores > 1 than parallisation for bootstrapping is initialized via future backend. Default: 1}
}
\value{
vector containing rainfall metrics (see description). If return.DataFrame is TRUE a data.frame is returned containing similar rain metrics for all rain events.
}
\description{
This function calcuates different precipitation characteristics for a specific time-series:
total precipitation, number of rainfall events, weighted mean intensitiy of rainfall events (normalized by MAP, RD or RDN),
cumulative critical event rainfall (normalized by MAP, RD or RDN), maximum rainfall during critical rainfall event,
duration of critical rainfall event, critical rainfall intensitiy (normalized by MAP, RD or RDN), rainfall at day of failure (start date),
rainfall intensity at day of failure (start date), maximum rainfall at day of failure (start date).
}
\note{
The use of method = "NLS" is not recommended.

\itemize{
\item Brunetti, M. T., Peruccacci, S., Rossi, M., Luciani, S., Valigi, D., & Guzzetti, F. (2010). Rainfall thresholds for the possible occurrence of landslides in Italy. Natural Hazards and Earth System Sciences, 10(3), 447.
\item Peruccacci, S., Brunetti, M. T., Luciani, S., Vennari, C., & Guzzetti, F. (2012). Lithological and seasonal control on rainfall thresholds for the possible initiation of landslides in central Italy. Geomorphology, 139, 79-90.
\item Rossi, M., Luciani, S., Valigi, D., Kirschbaum, D., Brunetti, M. T., Peruccacci, S., & Guzzetti, F. (2017). Statistical approaches for the definition of landslide rainfall thresholds and their uncertainty using rain gauge and satellite data. Geomorphology, 285, 16-27.
\item Guzzetti, F., Peruccacci, S., Rossi, M., & Stark, C. P. (2007). Rainfall thresholds for the initiation of landslides in central and southern Europe. Meteorology and atmospheric physics, 98(3-4), 239-267.
}
}
\keyword{appraoch}
\keyword{automatic}
\keyword{event,}
\keyword{landslide,}
\keyword{rainfall}
\keyword{tresholds,}
